move related files into catalog module
