+++ frontend/src/modules/Catalog/view/GlossarySearch.js
@@ -6,10 +6,10 @@ import { makeStyles } from '@mui/styles';
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import * as BsIcons from 'react-icons/bs';
-import { Defaults, Scrollbar } from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
-import { searchGlossary, useClient } from '../../../../services';
-import { listToTree } from '../../../../utils';
+import { Defaults, Scrollbar } from '../../../design';
+import { SET_ERROR, useDispatch } from '../../../globalErrors';
+import { searchGlossary, useClient } from '../../../services';
+import { listToTree } from '../../../utils';
 
 const useTreeItemStyles = makeStyles((theme) => ({
   root: {
