+++ frontend/src/modules/Catalog/view/GlossarySearchComponent.js
