+++ frontend/src/modules/Catalog/view/index.js
